+++ .gitignore
@@ -1,4 +1,5 @@
 .pio
+.vscode/
 .vscode/.browse.c_cpp.db*
 .vscode/c_cpp_properties.json
 .vscode/launch.json
