+++ .pre-commit-config.yaml
@@ -3,7 +3,9 @@ repos:
     hooks:
       - id: run-unit-tests
         name: Run MinUnit tests
-        entry: sh -c "cd test && gcc -I. -o test_player_now test_player.c player_action.c && sleep 0.2 && ./test_player_now"
+        # entry: sh -c "cd test && gcc -I. -o test_player_now test_player.c player_action.c && sleep 0.2 && ./test_player_now"
+        entry: sh -c "gcc -I . -I include -o test_player_now ./test/test_player.c ./src/player_action.c && sleep 0.2 && ./test_player_now"
+
         language: system
         stages: [pre-commit]
       - id: cppcheck
@@ -15,7 +17,7 @@ repos:
         stages: [pre-commit]
       - id: run-coverage
         name: Run Coverage with GCC
-        entry: sh -c 'gcc -Iinclude -fprofile-arcs -ftest-coverage -o test_coverage test/*.c && ./test_coverage && mkdir -p cov_html && gcovr -r . --exclude test/test_player.c --exclude test/minunit.h --html --html-details -o cov_html/index.html && python -m webbrowser -t "cov_html/index.html"'
+        entry: sh -c 'gcc -Iinclude -fprofile-arcs -ftest-coverage -o test_coverage ./test/*.c ./src/player_action.c && ./test_coverage && mkdir -p cov_html && gcovr -r . --exclude test/test_player.c --exclude test/minunit.h --html --html-details -o cov_html/index.html && cd cov_html && python -m webbrowser -t "index.html"'
         language: system
         files: ^test/.*\.c$
         stages: [pre-commit]
